+++ es6.js
@@ -1,8 +1,0 @@
-module.exports = {
-  extends: [
-    "./legacy",
-    "./rules/es6",
-    "./rules/es6-plugins"
-  ].map(require.resolve),
-  rules: {}
-};