+++ rules/4-6.js
@@ -1,0 +1,6 @@
+module.exports = {
+  "rules": {
+    // enforce consistent line breaks inside function parentheses
+    "function-paren-newline": ["off"]
+  }
+};